chore: add skeleton to staking page
